requirements.txt
py-cord>=2.6.0
bilibili-api-python
yt-dlp
aiohttp
httpx
APScheduler
PyNaCl
requests
openai
